requirements/requirements.txt
cryptography==46.0.3
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
python-dotenv==1.2.1

